backend/requirements.txt
asgiref==3.8.1
boto3==1.38.36
botocore==1.38.36
CacheControl==0.14.3
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
iso3166==2.1.1
jmespath==1.0.1
mapbox==0.18.1
msgpack==1.1.1
polyline==2.0.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.4
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
